Add '--content-type auto' flag for object storage
Currently, we have to manually specify the content type of files we upload to an object storage bucket. This can become tedious when we do a bulk-upload because the content-type flag will upload all the files with the same content-type regardless of whether it is accurate. I propose using the mimetypes standard library in Python  and adding a '--content-type auto' flag when using object storage bulk-upload and put commands to automatically generate the content-type for each file. Apologies if a feature like this has been considered already. Thank you!